fix: update hero section layout, enhance site header link structure, and streamline page animations
